+++ src/main/java/courgette/integration/reportportal/ReportPortalService.java
@@ -1,191 +1,193 @@
 package courgette.integration.reportportal;
 
-import com.google.gson.JsonArray;
 import com.google.gson.JsonElement;
-import com.google.gson.JsonNull;
-import com.google.gson.JsonObject;
 import com.google.gson.JsonParser;
-import com.google.gson.JsonPrimitive;
+import courgette.integration.reportportal.request.FinishRequest;
+import courgette.integration.reportportal.request.LaunchRequest;
+import courgette.integration.reportportal.request.ScenarioLogRequest;
+import courgette.integration.reportportal.request.ScenarioRequest;
+import courgette.integration.reportportal.request.TestRequest;
+import courgette.integration.reportportal.request.TestSuiteRequest;
+import courgette.runtime.CourgetteProperties;
+import courgette.runtime.CourgetteRunnerInfo;
+import courgette.runtime.report.JsonReportParser;
+import courgette.runtime.report.model.Scenario;
+import courgette.runtime.report.model.Tag;
 import courgette.runtime.utils.FileUtils;
+import io.cucumber.core.gherkin.Feature;
+import org.apache.http.Header;
 import org.apache.http.HttpEntity;
 import org.apache.http.HttpResponse;
-import org.apache.http.HttpStatus;
 import org.apache.http.client.HttpClient;
-import org.apache.http.client.entity.EntityBuilder;
-import org.apache.http.client.methods.HttpGet;
 import org.apache.http.client.methods.HttpPost;
 import org.apache.http.client.methods.HttpPut;
+import org.apache.http.client.methods.HttpUriRequest;
 import org.apache.http.conn.ssl.TrustAllStrategy;
-import org.apache.http.entity.ContentType;
-import org.apache.http.entity.mime.MultipartEntityBuilder;
 import org.apache.http.impl.client.HttpClientBuilder;
+import org.apache.http.message.BasicHeader;
 import org.apache.http.ssl.SSLContextBuilder;
 import org.apache.http.util.EntityUtils;
 
 import javax.net.ssl.SSLContext;
 import java.io.File;
 import java.io.IOException;
-import java.util.Optional;
+import java.util.List;
+import java.util.concurrent.ConcurrentHashMap;
+import java.util.function.Function;
+import java.util.function.Predicate;
+import java.util.stream.Collectors;
 
-import static java.util.Arrays.asList;
+import static courgette.runtime.CourgetteException.printError;
 
 public class ReportPortalService {
+
     private final ReportPortalProperties reportPortalProperties;
+    private final CourgetteProperties courgetteProperties;
     private final HttpClient httpClient;
-    private final String authorization;
-    private final String importUrl;
-    private final String latestLaunchUrl;
+    private final List<Feature> features;
 
-    private final String API_LAUNCH_IMPORT = "api/v1/%s/launch/import";
-    private final String API_LAUNCH_LATEST = "api/v1/%s/launch/latest";
-    private final String API_LAUNCH_UPDATE = "api/v1/%s/launch/%d/update";
+    private final ConcurrentHashMap<String, String> testIds = new ConcurrentHashMap<>();
+    private final ConcurrentHashMap<String, Long> testCounters = new ConcurrentHashMap<>();
 
-    private Integer launchId;
+    private String launchId;
+    private String testSuiteId;
 
-    private ReportPortalService(ReportPortalProperties reportPortalProperties) {
+    public ReportPortalService(ReportPortalProperties reportPortalProperties,
+                               CourgetteProperties courgetteProperties,
+                               List<Feature> features) {
+
         this.reportPortalProperties = reportPortalProperties;
+        this.courgetteProperties = courgetteProperties;
         this.httpClient = createHttpClient();
-        this.authorization = "bearer " + reportPortalProperties.getApiToken();
-        this.importUrl = reportPortalProperties.getEndpoint() + String.format(API_LAUNCH_IMPORT, reportPortalProperties.getProject());
-        this.latestLaunchUrl = reportPortalProperties.getEndpoint() + String.format(API_LAUNCH_LATEST, reportPortalProperties.getProject());
+        this.features = features;
     }
 
-    public static ReportPortalService create(ReportPortalProperties reportPortalProperties) {
-        return new ReportPortalService(reportPortalProperties);
+    public void startLaunch() {
+        launchId = call(post(apiEndpoint() + "/launch", new LaunchRequest().create(reportPortalProperties)));
     }
 
-    public boolean publishReport(String reportFilename) {
-        File zipFile = FileUtils.zipFile(reportFilename, true);
+    public void startTestSuite() {
+        if (launchId != null) {
+            testSuiteId = call(post(apiEndpoint() + "/item", new TestSuiteRequest().create(reportPortalProperties.getTestSuite(), launchId)));
+        }
 
-        boolean published = false;
+        if (testSuiteId != null) {
+            createTestIds(testIds, features);
+            createTestCounters(testCounters, features);
+        }
+    }
 
-        if (zipFile.exists()) {
-            final Optional<HttpResponse> response = sendMultiPartPost(zipFile);
-            published = handleOKResponse(response);
+    public void addTest(CourgetteRunnerInfo runnerInfo) {
 
-            if (published) {
-                setLatestLaunchId();
+        if (testSuiteId != null) {
+
+            final String testName = createFeatureName(runnerInfo.getFeature());
+
+            final File reportFile = runnerInfo.getJsonReportFile();
+
+            if (reportFile == null) {
+                printError("Courgette Report Portal Service: unable to add test due to missing report data. Feature URI -> " + runnerInfo.getFeature().getUri());
+                return;
+            }
+
+            final List<courgette.runtime.report.model.Feature> features = JsonReportParser
+                    .create(reportFile, courgetteProperties.getCourgetteOptions().runLevel())
+                    .getReportFeatures();
+
+            final String testId = testIds.get(testName);
+
+            List<Scenario> scenarios = features.stream()
+                    .flatMap(feature -> feature.getScenarios().stream())
+                    .collect(Collectors.toList());
+
+            for (Scenario scenario : scenarios) {
+                if (testId != null) {
+                    final String scenarioId = startScenario(testId, scenario);
+
+                    if (scenarioId != null) {
+                        addScenarioLog(scenarioId, scenario);
+                        finishScenario(scenarioId, scenario.passed() ? "passed" : "failed");
+                        testCounters.put(testName, testCounters.get(testName) - 1);
+                    }
+                }
+            }
+
+            if (testCounters.get(testName) <= 0) {
+                finishTest(testId);
             }
         }
-        return published;
     }
 
-    public void updateLaunchTags() {
-        final String tags = reportPortalProperties.getTags();
-
-        if (launchId != null && !tags.isEmpty()) {
-            String launchUpdateUrl = reportPortalProperties.getEndpoint() + String.format(API_LAUNCH_UPDATE, reportPortalProperties.getProject(), launchId);
-
-            String updateRequest = createLaunchUpdateRequest(tags);
-            final Optional<HttpResponse> response = updateLaunchTags(launchUpdateUrl, updateRequest);
-            handleOKResponse(response);
+    public void finishTestSuite() {
+        if (testSuiteId != null) {
+            call(put(apiEndpoint() + "/item/" + testSuiteId, new FinishRequest().create(launchId)));
         }
     }
 
-    private Optional<HttpResponse> sendMultiPartPost(File file) {
-        try {
-            HttpEntity entity = MultipartEntityBuilder
-                    .create()
-                    .addBinaryBody("file", file)
-                    .build();
-
-            HttpPost importReport = new HttpPost(importUrl);
-            importReport.addHeader("Authorization", authorization);
-            importReport.setEntity(entity);
-            return Optional.of(httpClient.execute(importReport));
-        } catch (Exception e) {
-            printError("publish the report", e.getMessage());
-        }
-        return Optional.empty();
-    }
-
-    private Optional<HttpResponse> updateLaunchTags(String launchUpdateUrl, String updateRequest) {
-        HttpPut launchUpdate = new HttpPut(launchUpdateUrl);
-        launchUpdate.addHeader("Authorization", authorization);
-
-        HttpEntity entity = EntityBuilder
-                .create()
-                .setContentType(ContentType.APPLICATION_JSON)
-                .setText(updateRequest)
-                .build();
-
-        launchUpdate.setEntity(entity);
-
-        try {
-            return Optional.ofNullable(httpClient.execute(launchUpdate));
-        } catch (Exception e) {
-            printError("update launch tags", e.getMessage());
-        }
-        return Optional.empty();
-    }
-
-    private void setLatestLaunchId() {
-        HttpGet latestLaunch = new HttpGet(latestLaunchUrl);
-        latestLaunch.addHeader("Authorization", authorization);
-
-        try {
-            HttpResponse response = httpClient.execute(latestLaunch);
-
-            if (response.getStatusLine().getStatusCode() == HttpStatus.SC_OK) {
-                String body = EntityUtils.toString(response.getEntity(), "UTF-8");
-                Optional<Integer> id = extractLaunchId(body);
-                id.ifPresent(i -> launchId = i);
-            }
-        } catch (Exception e) {
-            printError("get the latest launch id", e.getMessage());
+    public void finishLaunch() {
+        if (launchId != null) {
+            call(put(apiEndpoint() + "/launch/" + launchId + "/finish", new FinishRequest().create(launchId)));
         }
     }
 
-    private Optional<Integer> extractLaunchId(String body) {
-        JsonElement json = JsonParser.parseString(body);
-        if (json instanceof JsonNull) {
-            return Optional.empty();
-        }
-
-        JsonElement content = json.getAsJsonObject().get("content");
-        if (content instanceof JsonNull) {
-            return Optional.empty();
-        }
-
-        JsonArray contentDetails = content.getAsJsonArray();
-        JsonElement id = contentDetails.get(contentDetails.size() - 1).getAsJsonObject().get("id");
-        return (id instanceof JsonNull) ? Optional.empty() : Optional.of(id.getAsInt());
+    private String startTest(String featureName) {
+        return call(post(apiEndpoint() + "/item/" + testSuiteId, TestRequest.create(featureName, launchId)));
     }
 
-    private String createLaunchUpdateRequest(String tags) {
-        JsonObject jsonObject = new JsonObject();
-        JsonArray attributeArr = new JsonArray();
-
-        String[] tagData = tags.split(";");
-
-        asList(tagData).forEach(tag -> {
-            JsonObject tagObj = new JsonObject();
-            tagObj.add("key", JsonNull.INSTANCE);
-            tagObj.add("value", new JsonPrimitive(tag));
-            attributeArr.add(tagObj);
-        });
-
-        jsonObject.add("attributes", attributeArr);
-        jsonObject.add("mode", new JsonPrimitive("DEFAULT"));
-        return jsonObject.toString();
+    private void finishTest(String testId) {
+        if (testId != null) {
+            call(put(apiEndpoint() + "/item/" + testId, new FinishRequest().create(launchId)));
+        }
     }
 
-    private boolean handleOKResponse(Optional<HttpResponse> response) {
-        if (!response.isPresent()) {
-            return false;
-        }
+    private String startScenario(String testId, Scenario scenario) {
+        return call(post(apiEndpoint() + "/item/" + testId,
+                ScenarioRequest.create(
+                        scenario.getName(),
+                        scenario.getTags().stream().map(Tag::getName).collect(Collectors.toList()),
+                        launchId)));
+    }
 
-        if (response.get().getStatusLine().getStatusCode() != HttpStatus.SC_OK) {
-            String body;
-            try {
-                body = EntityUtils.toString(response.get().getEntity(), "UTF-8");
-            } catch (IOException e) {
-                body = e.getMessage();
-            }
-            printError(body);
-            return false;
-        }
-        return true;
+    private void addScenarioLog(String scenarioId, Scenario scenario) {
+        call(post(apiEndpoint() + "/log", new ScenarioLogRequest().create(scenario, scenarioId, launchId)));
+    }
+
+    private void finishScenario(String scenarioId, String status) {
+        call(put(apiEndpoint() + "/item/" + scenarioId, new FinishRequest().create(status, launchId)));
+    }
+
+    private String createFeatureName(Feature feature) {
+        return FileUtils.getFileName(feature.getUri()).split("\\.")[0];
+    }
+
+    private void createTestIds(ConcurrentHashMap<String, String> testIds, List<Feature> features) {
+        features.stream().map(this::createFeatureName).distinct().sorted().forEach(testName -> testIds.put(testName, startTest(testName)));
+    }
+
+    private void createTestCounters(ConcurrentHashMap<String, Long> testCounters, List<Feature> features) {
+        testCounters.putAll(features.stream().map(this::createFeatureName).collect(Collectors.groupingBy(Function.identity(), Collectors.counting())));
+    }
+
+    private String apiEndpoint() {
+        return String.format("%s/api/v1/%s", reportPortalProperties.getEndpoint(), reportPortalProperties.getProject());
+    }
+
+    private HttpPost post(final String uri, HttpEntity entity) {
+        HttpPost post = new HttpPost(uri);
+        post.addHeader(authHeader());
+        post.setEntity(entity);
+        return post;
+    }
+
+    private HttpPut put(final String uri, HttpEntity entity) {
+        HttpPut put = new HttpPut(uri);
+        put.addHeader(authHeader());
+        put.setEntity(entity);
+        return put;
+    }
+
+    private Header authHeader() {
+        return new BasicHeader("Authorization", "bearer " + reportPortalProperties.getApiToken());
     }
 
     private HttpClient createHttpClient() {
@@ -193,16 +195,38 @@ public class ReportPortalService {
             SSLContext trustedSSLContext = new SSLContextBuilder().loadTrustMaterial(null, TrustAllStrategy.INSTANCE).build();
             return HttpClientBuilder.create().setSSLContext(trustedSSLContext).build();
         } catch (Exception e) {
-            printError("error creating a secure http client: " + e.getMessage());
+            printError("Courgette Report Portal Service: error creating a secure http client: " + e.getMessage());
         }
+
         return HttpClientBuilder.create().build();
     }
 
-    private void printError(String error) {
-        System.err.format("Courgette Report Portal Service: %s\n", error);
+    private synchronized String call(final HttpUriRequest request) {
+        String id = null;
+
+        String responseBody;
+
+        try {
+            HttpResponse response = httpClient.execute(request);
+
+            responseBody = EntityUtils.toString(response.getEntity(), "UTF-8");
+
+            if (isSuccessfulCall.test(response.getStatusLine().getStatusCode())) {
+
+                JsonElement idElement = JsonParser.parseString(responseBody).getAsJsonObject().get("id");
+
+                if (idElement != null) {
+                    id = idElement.getAsString();
+                }
+            } else {
+                printError("Courgette Report Portal Service: " + responseBody);
+            }
+        } catch (IOException e) {
+            printError("Courgette Report Portal Service: " + e.getMessage());
+        }
+
+        return id;
     }
 
-    private void printError(String action, String error) {
-        System.err.format("Courgette Report Portal Service: unable to %s on the report portal server, reason: %s\n", action, error);
-    }
+    private final Predicate<Integer> isSuccessfulCall = (status) -> status == 200 || status == 201;
 }